refactor: ViewModelAutoInjection 의존성 분리
